requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests>=2.32.3
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
structlog==23.2.0
pydantic>2.5.0
pydantic-settings==2.1.0
tenacity==8.2.3
aiohttp==3.9.1
python-json-logger==2.0.7
cryptography==41.0.7
psutil==5.9.6
